% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactuals.R
\name{bfs_sfx}
\alias{bfs_sfx}
\title{Title}
\usage{
bfs_sfx(predict_f, evidence, target, expected)
}
\arguments{
\item{predict_f}{A function that takes in a dataframe of evidence and outputs
a list with the prediction for each row.}

\item{evidence}{A dataframe of one row where all columns are of type factor.}

\item{target}{The resulting prediction for the evidence.}

\item{expected}{The expected prediction of the evidence (only for
counterfactuals).}
}
\value{
List where each element is a sufficient explanation of the prediction.
}
\description{
Title
}
\examples{
library(bnlearn)
set.seed(40)
download.file("https://www.bnlearn.com/bnrepository/child/child.rda",
              "child.rda", "auto", quiet = TRUE)
load("child.rda")  # Load CHILD Bayesian Network
evidence = data.frame(
  LVHReport=factor(x = "yes", levels = dimnames(bn$LVHreport$prob)[[1]]),
  LowerBodyO2=factor(x = "5-12", levels = dimnames(bn$LowerBodyO2$prob)[[1]]),
  CO2Report=factor(x = "<7.5", levels = dimnames(bn$CO2Report$prob)[[1]]),
  XrayReport=factor(x = "Oligaemic", levels = dimnames(bn$XrayReport$prob)[[1]])
)
target <- predict(object = bn, node = bn$Disease$node, data = evidence,
                 method = "bayes-lw")
print(target)
expected <- factor("PAIVS", levels = levels(target))
predict_f <- function(df) {
  predict(object = bn, node = bn$Disease$node, data = df, method = "bayes-lw")}
print(bfs_sfx(predict_f, evidence, target, expected))
}
