% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactuals.R
\name{bnlearn_predict_wrapper}
\alias{bnlearn_predict_wrapper}
\title{Create a predict function for use of a bnlearn bayesian network in the
search algorithms.}
\usage{
bnlearn_predict_wrapper(
  bn,
  node,
  method,
  threshold = NULL,
  default = NULL,
  n = 500
)
}
\arguments{
\item{bn}{Fitted bayesian network}

\item{node}{Name of target node to be predicted}

\item{method}{Same options as in parameter 'method' in predict function of
bnlearn}

\item{threshold}{Minimum value of posterior probability of the MLE to be
the output of the algorithm. If parameter 'default' is null, the MLE will
be the output even if it doesn't surpass the threshold.}

\item{default}{Default predicted target value when threshold isn't exceeded
by the MLE}

\item{n}{Same options as in parameter 'n' in predict function of bnlearn}
}
\value{
Function to be used in algorithms such as bfs_sfx
}
\description{
Create a predict function for use of a bnlearn bayesian network in the
search algorithms.
}
\examples{
library(bnlearn)
set.seed(40)
download.file("https://www.bnlearn.com/bnrepository/child/child.rda",
              "child.rda", "auto", quiet = TRUE)
load("child.rda")  # Load CHILD Bayesian Network
evidence = data.frame(
  LVHReport=factor(x = "yes", levels = dimnames(bn$LVHreport$prob)[[1]]),
  LowerBodyO2=factor(x = "5-12", levels = dimnames(bn$LowerBodyO2$prob)[[1]]),
  CO2Report=factor(x = "<7.5", levels = dimnames(bn$CO2Report$prob)[[1]]),
  XrayReport=factor(x = "Oligaemic", levels = dimnames(bn$XrayReport$prob)[[1]])
)
target <- predict(object = bn, node = bn$Disease$node, data = evidence,
                 method = "bayes-lw")
expected <- factor("PAIVS", levels = levels(target))
predict_f <- bnlearn_predict_wrapper(bn, bn$Disease$node, "bayes-lw",
                                     threshold=0.20, default="TGA")
print(bfs_sfx(predict_f, evidence, target, expected))
}
